Divar posts dataset

The dataset contains 947635 posts that were published in Divar classified ads platform. These posts were published and archived before 2017. Distribution of these posts in different groups does not resemble the actual distributions. The dataset contains posts from six major categories in Divar:

Vehicles

Electronic devices

Businesses

For the home

Personal

Leisure & Hobbies

data(divar)

Format

A tibble with 947635 rows and 17 variables:

id

Unique id of the post (1246772 ids)

archive_by_user

Whether the post was archived by user or automatically by the system (True/False)

published_at

Weekday and hour the post was published

cat1

First level category of the post (for example cat1)

cat2

Second level category of the post (can be empty)

cat3

Third level category of the post (can be empty) (for example light)

city

Name of the city the ad was published in (for example Mashhad)

title

Title of the post. All phone numbers are replaced by the $NUM token.

city

Name of the city the ad was published in (for example Mashhad)

desc

Description of the post. All phone numbers are replaced by the $NUM token.

price

Name of the city the ad was published in (for example Mashhad)

image_count

Number of images for the post

platform

The platform the post was submitted from (mobile or web)

mileage

(Only for light vehicles) the mileage of the vehicle posted in kilometers.

brand

(Only for light vehicles and electronic devices) English and Persian name of the brand separated with ::(two colons)

year

(Only for light vehicles) production year of the vehicle in Iranian calendar. Can be the special value of <1366 which means older than year 1366.

type

(Only for clothing) boys/girls for children clothing and men/women for adult clothing

Source

Cafebazaar Research Group

Examples

divar

#> Error in eval(expr, envir, enclos): object 'divar' not found

if (require("dplyr")) {
divar %>%
  select(nutr_id, deriv_id) %>%
  left_join(deriv)
}

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#> Error in eval(lhs, parent, parent): object 'divar' not found