Spotify

A datas Iranian Artists on

spotify

Format

A tibble with 10632 rows and 32 variables:

species

a factor denoting penguin species (Adélie, Chinstrap and Gentoo)

island

a factor denoting island in Palmer Archipelago, Antarctica (Biscoe, Dream or Torgersen)

bill_length_mm

a number denoting bill length (millimeters)

bill_depth_mm

a number denoting bill depth (millimeters)

flipper_length_mm

an integer denoting flipper length (millimeters)

body_mass_g

an integer denoting body mass (grams)

sex

a factor denoting penguin sex (female, male)

year

an integer denoting the study year (2007, 2008, or 2009)

Source

Collected by Muhammad Chenariyan Nakhaee using Spotify's Web API and Spotifyr package) mcnakhaee.com

Details

The original data included cases where a country was absent or was not yet a member. In this dataset these were filtered out to include only votes of Yes, Abstain, and No.

Country name can be converted to other unique country identifiers (such as 2-character or 3-character ISO codes) using the countrycode function from the countrycode package.

Examples

library(tidyverse)

#> Error in library(tidyverse): there is no package called ‘tidyverse’