Ubaar Transport Cost Prediction (Test)

In this competition, Ubaar & SADA challenge you to predict transport cost for 15,000 loads. You are provided with data about 50,000 actual transported loads. You can learn more about this competition at http://sanags.ir/ubaar-competition

ubaar_test

Format

A tibble with 50000 rows and 14 variables:

ID

an anonymous id unique to a given load

date

the date of the load dispatch. It is a string with ymd format. for example, 960630 is 30th day of 6th month of year 1396.

sourceLatitude

latitude of the source region of transport. Source region is not exactly the location for embarkation of the freight.

sourceLongitude

longitude of the source region of transport. Source region is not exactly the location for embarkation of the freight.

SourceState

state of the source of transport

destinationLatitude

latitude of the destination region of transport. Destination region is not exactly the location for disembarkation of the freight.

destinationLongitude

latitude of the destination region of transport. Destination region is not exactly the location for disembarkation of the freight.

destinationState

state of the destination of transport

distanceKM

driving distance in kilo meters from source region to destination region for usual cars, calculated using google distance services

taxiDurationMin

estimated arrival time in minutes from source region to destination region for usual cars, calculated using google distance services

vehicleType

type of the vehicle that has carried the load. It is a categorical variable

vehicleOption

option of the vehicle that has carried the load. It is a categorical variable. Each vehicle type has its own vehicle options

weight

weight of the load that has been carried in Tons (1000KG)

Source

Ubaar Kaggle Compettion